Update Gradle wrapper and build script
Based on the newest version of "carttemplate" + some more modern Loom & Gradle conventions. Using the diff command to compare unzipped builds, I'm fairly sure that nothing is broken. More worrying is why I can't reproduce the original builds of the mod. Not sure how to remove the refmap shenanigans in this script yet, Loom still has issues with this apparently.